feat: add `ModelXmlSerializer` to import/export the model using XML
This class was previously only available in test. It simplifies maxGraph model import/export as:
  - it hides `Codec` and `HTMLElement` management
  - it registers codecs automagically
